remove ScriptedImporterExtension
