+++ Assets/UniGLTF/Editor/UniGLTF/ScriptedImporter/RemapEditorAnimation.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using UnityEditor;
 using UnityEditor.Experimental.AssetImporters;
@@ -35,16 +32,6 @@ namespace UniGLTF
             DrawRemapGUI<AnimationClip>(importer.GetExternalObjectMap());
         }
 
-        static string GetAndCreateFolder(string assetPath, string suffix)
-        {
-            var path = $"{Path.GetDirectoryName(assetPath)}/{Path.GetFileNameWithoutExtension(assetPath)}{suffix}";
-            if (!Directory.Exists(path))
-            {
-                Directory.CreateDirectory(path);
-            }
-            return path;
-        }
-
         public static void Extract(ScriptedImporter importer, GltfParser parser)
         {
             if (string.IsNullOrEmpty(importer.assetPath))
@@ -52,12 +39,12 @@ namespace UniGLTF
                 return;
             }
 
-
+            var path = GetAndCreateFolder(importer.assetPath, ".Animations");
+            foreach (var asset in AssetDatabase.LoadAllAssetsAtPath(importer.assetPath))
             {
-                var path = GetAndCreateFolder(importer.assetPath, ".Animations");
-                foreach (var (key, asset) in importer.GetSubAssets<AnimationClip>(importer.assetPath))
+                if (asset is AnimationClip)
                 {
-                    asset.ExtractSubAsset($"{path}/{asset.name}.asset", false);
+                    ExtractSubAsset(asset, $"{path}/{asset.name}.asset", false);
                 }
             }
