+++ Assets/UniGLTF/Editor/UniGLTF/ScriptedImporter/RemapEditorBase.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using UnityEditor;
 using UnityEditor.Experimental.AssetImporters;
@@ -99,5 +100,64 @@ namespace UniGLTF
             }
             EditorGUI.indentLevel--;
         }
+
+        protected static string GetAndCreateFolder(string assetPath, string suffix)
+        {
+            var path = $"{Path.GetDirectoryName(assetPath)}/{Path.GetFileNameWithoutExtension(assetPath)}{suffix}";
+            if (!Directory.Exists(path))
+            {
+                Directory.CreateDirectory(path);
+            }
+            return path;
+        }
+
+        /// <summary>
+        /// subAsset を 指定された path に extract する
+        /// </summary>
+        /// <param name="subAsset"></param>
+        /// <param name="destinationPath"></param>
+        /// <param name="isForceUpdate"></param>
+        public static UnityEngine.Object ExtractSubAsset(UnityEngine.Object subAsset, string destinationPath, bool isForceUpdate)
+        {
+            string assetPath = AssetDatabase.GetAssetPath(subAsset);
+
+            // clone を path に出力(subAsset を出力するため)
+            var clone = UnityEngine.Object.Instantiate(subAsset);
+            AssetDatabase.CreateAsset(clone, destinationPath);
+
+            // subAsset を clone に対して remap する
+            var assetImporter = AssetImporter.GetAtPath(assetPath);
+            assetImporter.AddRemap(new AssetImporter.SourceAssetIdentifier(clone), clone);
+
+            if (isForceUpdate)
+            {
+                AssetDatabase.WriteImportSettingsIfDirty(assetPath);
+                AssetDatabase.ImportAsset(assetPath, ImportAssetOptions.ForceUpdate);
+            }
+
+            return clone;
+        }
+
+        // public static void ClearExternalObjects(this ScriptedImporter importer, params Type[] targetTypes)
+        // {
+        //     foreach (var targetType in targetTypes)
+        //     {
+        //         if (!typeof(UnityEngine.Object).IsAssignableFrom(targetType))
+        //         {
+        //             throw new NotImplementedException();
+        //         }
+
+        //         foreach (var (key, obj) in importer.GetExternalObjectMap())
+        //         {
+        //             if (targetType.IsAssignableFrom(key.type))
+        //             {
+        //                 importer.RemoveRemap(key);
+        //             }
+        //         }
+        //     }
+
+        //     AssetDatabase.WriteImportSettingsIfDirty(importer.assetPath);
+        //     AssetDatabase.ImportAsset(importer.assetPath, ImportAssetOptions.ForceUpdate);
+        // }
     }
 }
